Improve net.uri queries in pyvast-threatbus
Substring queries in VAST are slow compared to point queries. Instead
of using `in` to query for the concept `net.uri`, we use strict equality
in case the indicator begins with a protocol identifier, and otherwise
search for the indicator and the indicator prefixed with common prefixes
like `http://` or `tcp://` instead.